% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_ff_data.R
\name{get_ff_data}
\alias{get_ff_data}
\title{function for getting data on follower and followings}
\usage{
get_ff_data(screen_name)
}
\arguments{
\item{screen_name}{name of the twitter account to be analyzed which is passed to twitteR::getUser()}
}
\description{
the function retrieves name, id, screen_name of followers (those who listen to you) and following (those you listen to) as well as friends (those you listen to and which listen back and vice versa)
}

