% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_twitter_log.R
\name{get_twitter_log}
\alias{get_twitter_log}
\title{function for retrieving twitter log from log-file}
\usage{
get_twitter_log(fname)
}
\description{
function for retrieving twitter log from log-file
}

