% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.twitterdat.R
\name{plot.twitterdat}
\alias{plot.twitterdat}
\title{plot method for twitterdat}
\usage{
\method{plot}{twitterdat}(x, ...)
}
\arguments{
\item{x}{data to be plotted}
}
\description{
plot method for twitterdat
}

