% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twitter_log.R
\name{twitter_log}
\alias{twitter_log}
\title{logging twitter statistics}
\usage{
twitter_log(screen_name, file = T)
}
\arguments{
\item{screen_name}{name or id for which statistics should be gathered}

\item{file}{wether or not info should be added to log file, might be either T/F or the name of the log-file; if set to true the log file name will consists of the account namewit hending 'twitterlog'}
}
\description{
logging twitter statistics
}

